@@ -4,7 +4,7 @@ description: Demonstrates the Windows implementation of the url_launcher plugin.
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher_platform_interface: any
+  url_launcher_platform_interface: ^2.0.0
   url_launcher_windows:
     # When depending on this package from a real application you should use:
     #   url_launcher_windows: ^x.y.z
@@ -18,11 +18,11 @@ dev_dependencies:
     path: ../../../integration_test
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.8"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
