@@ -4,11 +4,11 @@
 
 import 'dart:async';
 
+import 'package:battery/battery.dart';
 import 'package:battery_platform_interface/battery_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:battery/battery.dart';
-import 'package:mockito/mockito.dart';
+import 'package:test/fake.dart';
 
 void main() {
   group('battery', () {
@@ -30,7 +30,7 @@ void main() {
   });
 }
 
-class MockBatteryPlatform extends Mock
+class MockBatteryPlatform extends Fake
     with MockPlatformInterfaceMixin
     implements BatteryPlatform {
   Future<int> batteryLevel() async {
