@@ -1068,7 +1068,7 @@ NSString *const errorMethod = @"error";
     result(getFlutterError(error));
     return;
   }
-  [_captureDevice rampToVideoZoomFactor:zoom withRate:1];
+  _captureDevice.videoZoomFactor = zoom;
   [_captureDevice unlockForConfiguration];
 
   result(nil);
